requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
google-generativeai==0.3.2
requests==2.31.0
streamlit-option-menu==0.3.12
plotly==5.19.0
pandas==2.2.1
Pillow==10.2.0
fpdf
beautifulsoup4
selenium
webdriver-manager
replicate
pyperclip
reportlab

